+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- exit and log out confirmation
+- logged-in NIK label
+
 ## [0.8.0] - 2020-05-27
 
 - setting functionality